requirements.txt
black==19.10b0
pandas==0.25.3
pg8000==1.13.2
pre-commit==1.20.0
pytest==5.2.2
pytest-cov==2.8.1
python-dotenv==0.10.3
SQLAlchemy==1.3.10
xlrd==1.2.0
